Refactor Code : Separate JobsController to JobSearchController and JobStatisticController
