+++ src/main/java/com/techlooper/controller/JobStatisticController.java
@@ -1,0 +1,58 @@
+package com.techlooper.controller;
+
+import com.techlooper.model.*;
+import com.techlooper.service.JobStatisticService;
+import org.springframework.messaging.handler.annotation.MessageMapping;
+import org.springframework.messaging.handler.annotation.SendTo;
+import org.springframework.messaging.simp.SimpMessagingTemplate;
+import org.springframework.scheduling.annotation.Scheduled;
+import org.springframework.stereotype.Controller;
+
+import javax.annotation.Resource;
+import java.util.Arrays;
+import java.util.LinkedList;
+import java.util.List;
+
+@Controller
+public class JobStatisticController {
+
+    @Resource
+    private JobStatisticService vietnamWorksJobStatisticService;
+
+    @Resource
+    private SimpMessagingTemplate messagingTemplate;
+
+    @Scheduled(cron = "${scheduled.cron}")
+    public void countTechnicalJobs() {
+        Arrays.stream(TechnicalTermEnum.values()).forEach(term ->
+                        messagingTemplate.convertAndSend("/topic/jobs/term/" + term.name(), new JobStatisticResponse.Builder()
+                                .withCount(vietnamWorksJobStatisticService.count(term)).build())
+        );
+    }
+
+    @SendTo("/topic/jobs/terms")
+    @MessageMapping("/jobs/terms")
+    public List<TechnicalTermResponse> countTechnicalTerms() {
+        List<TechnicalTermResponse> terms = new LinkedList<>();
+        Arrays.stream(TechnicalTermEnum.values()).forEach(term ->
+                        terms.add(new TechnicalTermResponse.Builder().withTerm(term)
+                                .withCount(vietnamWorksJobStatisticService.count(term)).build())
+        );
+        return terms;
+    }
+
+    @MessageMapping("/jobs")
+    public void countTechnicalJobs(JobStatisticRequest request) {
+        messagingTemplate.convertAndSend(
+                "/topic/jobs/" + request.getTerm().toLowerCase(),
+                new JobStatisticResponse.Builder().withCount(
+                        vietnamWorksJobStatisticService.count(TechnicalTermEnum.valueOf(request.getTerm().toUpperCase())))
+                        .build());
+    }
+
+    @SendTo("/topic/analytics/skill")
+    @MessageMapping("/analytics/skill")
+    public SkillStatisticResponse countTechnicalSkillByTerm(SkillStatisticRequest skillStatisticRequest) {
+        return vietnamWorksJobStatisticService.countJobsBySkill(skillStatisticRequest.getTerm(), skillStatisticRequest.getHistograms());
+    }
+}
